命名を変更してショップ画面のガワを作成
